+++ src/main/client/index.ts
@@ -1,4 +1,5 @@
 import './startup.js';
+
 import { useClientApi } from './api/index.js';
 
 import { useClientInteraction } from './controllers/interaction.js';